pyproject.toml
[tool.poetry]
name = "ai-agent-trade-claude"
version = "0.1.0"
description = "LangGraph-based Trading Agent System"
authors = ["Kevin"]
readme = "README.md"
packages = [{include = "app", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
langgraph = "^0.2.0"
langchain = "^0.2.0"
langchain-core = "^0.2.0"
# Gemini integration via direct SDK
google-generativeai = "^0.3.0"
pydantic = "^2.0.0"
pydantic-settings = "^2.0.0"
aiohttp = "^3.9.0"
websockets = "^12.0"
uvicorn = "^0.27.0"
fastapi = "^0.109.0"
python-binance = "^1.0.19"
# Alpaca paper trading integration
alpaca-py = "^0.20.0"
pandas = "^2.1.0"
numpy = "^1.26.0"
python-dotenv = "^1.0.0"
rich = "^13.0.0"
tenacity = "^8.2.0"
pandas-ta = "^0.3.14b"
statsmodels = "^0.14.0"
arch = "^6.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
black = "^23.12.0"
ruff = "^0.1.9"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["src/tests"]
